refactor: remove duplication between helper fns
